clearcut_detection_backend/requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
Django==2.2.3
djangorestframework==3.10.1
Fiona==1.8.6
geopandas==0.5.1
idna==2.8
munch==2.3.2
numpy==1.16.4
pandas==0.24.2
psycopg2-binary==2.8.3
pyproj==2.2.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
Shapely==1.6.4.post2
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
